Move admin comment handlers to handler package
