+++ server/handler/comment.go
@@ -49,6 +49,7 @@ func CommentAdd(c *gin.Context) {
 	c.JSON(http.StatusOK, model.RespOk("发布成功，返回评论ID", cmt.Id))
 }
 
+// 获取评论列表
 func CommentList(c *gin.Context) {
 	role, userId := utils.GetUserInfo(c)
 	condition := parseCommentWhere(c)
@@ -83,6 +84,7 @@ func CommentRemove(c *gin.Context) {
 	c.JSON(http.StatusOK, model.RespOk("删除成功", nil))
 }
 
+// 解析评论查询条件
 func parseCommentWhere(c *gin.Context) dao.CommentWhere {
 	condition := dao.CommentWhere{}
 
@@ -134,4 +136,43 @@ func parseCommentWhere(c *gin.Context) dao.CommentWhere {
 		}
 	}
 	return condition
+}
+
+// 修改评论
+type ReqCommentModify struct {
+	Id      uint64               `json:"id,omitempty" binding:"required"`
+	UserId  uint64               `json:"user_id,omitempty" binding:"required"`
+	BlogId  uint64               `json:"blog_id,omitempty" binding:"required"`
+	Content string               `json:"content,omitempty" binding:"required"`
+	Status  entity.CommentStatus `json:"status,omitempty"`
+}
+
+func CommentModify(c *gin.Context) {
+	var req ReqCommentModify
+
+	// 参数绑定
+	err := c.ShouldBindBodyWithJSON(&req)
+	if err != nil {
+		log.Println(err)
+		c.JSON(http.StatusBadRequest, model.RespOk("参数错误", nil))
+		return
+	}
+
+	cmt := entity.Comment{
+		Id:      req.Id,
+		UserId:  req.UserId,
+		BlogId:  req.BlogId,
+		Content: req.Content,
+		Status:  req.Status,
+	}
+
+	// 修改评论
+	err = comment.UpdateById(cmt)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, model.RespOk(err.Error(), nil))
+		return
+	}
+
+	// 返回结果
+	c.JSON(http.StatusOK, model.RespOk("修改成功", nil))
 }
